pyproject.toml
[project]
name = "health-wellness-agent"
version = "0.1.0"
description = "AI-powered Health & Wellness Planner using OpenAI Agents SDK"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai-agents>=0.1.0",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "streamlit>=1.46.1",
]
